Add Docker ports
